requirements.txt
annoy==1.17.3
bidict==0.23.1
boto3==1.42.6
cohere==5.20.0
deep_sort_realtime==1.3.2
dsntnn==0.5.3
fastapi_poe==0.0.81
librosa==0.11.0
matplotlib==3.10.8
networkx==3.2.1
numpy==2.3.5
openai==2.9.0
opencv_python==4.12.0.88
paramiko==4.0.0
Pillow==12.0.0
pydub==0.25.1
PyYAML==6.0.3
Requests==2.32.5
rich==14.2.0
rsa==4.9.1
scikit_learn==1.8.0
scipy==1.16.3
skimage==0.0
sshtunnel==0.4.0
thop==0.1.1.post2209072238
torch==2.2.2+cu118
torchaudio==2.2.2+cu118
torchvideotransforms==0.1.2
torchvision==0.17.2+cu118
tqdm==4.67.1
ultralytics==8.3.235
ultralytics_thop==2.0.18
volcengine_python_sdk==4.0.41
